+++ core/src/com/coronadefense/receiver/messages/PathToPathAnimationMessage.kt
@@ -2,9 +2,9 @@ package com.coronadefense.receiver.messages
 
 data class PathToPathAnimationMessage(
     val spriteNumber: UByte,
-    val startPosition: UShort,
-    val endPosition: UShort,
-    val startTime: UShort,
-    val endTime: UShort,
-    val resultAnimation: UByte
+    val startPosition: Float,
+    val endPosition: Float,
+    val startTime: Float,
+    val endTime: Float,
+    val resultAnimation: UByte,
 ): IMessage
